+++ src/cmd/template.rs
@@ -124,6 +124,7 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), std::io::Error> {
     let mut opt = HashMap::new();
 
     opt.insert("se", if args.is_present("se") { "1" } else { "0" });
+    opt.insert("parallel", args.value_of("parallel").unwrap());
 
     opt.insert("fastqc", if args.is_present("fastqc") { "1" } else { "0" });
     opt.insert(
@@ -148,6 +149,10 @@ pub fn execute(args: &ArgMatches) -> std::result::Result<(), std::io::Error> {
         gen_fastqc(&context)?;
     }
 
+    if args.is_present("kmergenie") {
+        gen_kmergenie(&context)?;
+    }
+
     Ok(())
 }
 
@@ -159,6 +164,23 @@ fn gen_fastqc(context: &Context) -> std::result::Result<(), std::io::Error> {
     tera.add_raw_templates(vec![
         ("header", include_str!("../../templates/header.tera.sh")),
         ("template", include_str!("../../templates/2_fastqc.tera.sh")),
+    ])
+    .unwrap();
+
+    let rendered = tera.render("template", &context).unwrap();
+    intspan::write_lines(outname, &vec![rendered.as_str()])?;
+
+    Ok(())
+}
+
+fn gen_kmergenie(context: &Context) -> std::result::Result<(), std::io::Error> {
+    let outname = "2_kmergenie.sh";
+    eprintln!("Create {}", outname);
+
+    let mut tera = Tera::default();
+    tera.add_raw_templates(vec![
+        ("header", include_str!("../../templates/header.tera.sh")),
+        ("template", include_str!("../../templates/2_kmergenie.tera.sh")),
     ])
     .unwrap();