+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased - ReleaseDate
 
+* Add `Anchr template`
+
 ## 0.1.0 - 2020-10-23
 
 * Add `Anchr merge`